+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v4.0.15 (2025-05-07)
+
+### Fix
+
+- fixed processing of labels in extract
+
+### Refactor
+
+- update toolbox_utils and ruff refactor
+
 ## v4.0.14 (2024-06-13)
 
 ### Fix